Create RedditScan.py
A tool to 'scan down' any new submissions to Reddit by users that are to external sites, and if they pass a series of tests, cross-post them with the original user-submitted title to another subreddit. This can be used as a cross-posting bot or to carefully and methodically curate your own personal reading lists from a variety of subreddits, to ensure that you never miss any content.